Refactor fetch assets
